+++ src/Render/Entity_DrawSetup.cpp
@@ -172,7 +172,7 @@ void __cdecl FUN_00504b50(int param_1, int param_2, float *param_3,
         *(float *)((int)this_ + 0x4c) = 0.5f;
         *(float *)((int)this_ + 0x50) = 1.5f;
         *(char *)((int)this_ + 0x88) = 0;
-        FUN_00441e00(this_, 2, *(float *)(param_1 + 0x168), *(float *)(param_1 + 100),
+        FUN_00441e00(this_, 2, *(float *)(param_1 + 0x168), *(int *)(param_1 + 100),
                      *(float *)(param_1 + 0x68), *(float *)(param_1 + 0x6c),
                      *(float *)(param_1 + 0x70), *(float *)(param_1 + 0x58), 1170);
         *(char *)((int)this_ + 0x88) = -1;
@@ -198,10 +198,10 @@ void __cdecl FUN_00504b50(int param_1, int param_2, float *param_3,
             *(float *)((int)this_ + 0x4c) = 0.8f;
             *(float *)((int)this_ + 0x50) = 0.4f;
         }
-        FUN_00441e00(this_, 8, *(float *)(param_1 + 0x168), *(float *)(param_1 + 100),
+        FUN_00441e00(this_, 8, *(float *)(param_1 + 0x168), *(int *)(param_1 + 100),
                      *(float *)(param_1 + 0x68), *(float *)(param_1 + 0x6c),
                      *(float *)(param_1 + 0x70), *(float *)(param_1 + 0x58), 1171);
-        FUN_00441e00(this_, 0x44, *(float *)(param_1 + 0x168), *(float *)(param_1 + 100),
+        FUN_00441e00(this_, 0x44, *(float *)(param_1 + 0x168), *(int *)(param_1 + 100),
                      *(float *)(param_1 + 0x68), *(float *)(param_1 + 0x6c),
                      *(float *)(param_1 + 0x70), *(float *)(param_1 + 0x58), 1171);
         return;
@@ -210,10 +210,10 @@ void __cdecl FUN_00504b50(int param_1, int param_2, float *param_3,
         *(float *)((int)this_ + 0x48) = 1.0f;
         *(float *)((int)this_ + 0x4c) = 1.0f;
         *(float *)((int)this_ + 0x50) = 1.0f;
-        FUN_00441e00(this_, 2, *(float *)(param_1 + 0x168), *(float *)(param_1 + 100),
+        FUN_00441e00(this_, 2, *(float *)(param_1 + 0x168), *(int *)(param_1 + 100),
                      *(float *)(param_1 + 0x68), *(float *)(param_1 + 0x6c),
                      *(float *)(param_1 + 0x70), -1.0f, -1);
-        FUN_00441e00(this_, 0x44, 0.5f, *(float *)(param_1 + 100),
+        FUN_00441e00(this_, 0x44, 0.5f, *(int *)(param_1 + 100),
                      *(float *)(param_1 + 0x68), *(float *)(param_1 + 0x6c),
                      *(float *)(param_1 + 0x70), -1.0f, 1171);
         return;
@@ -229,14 +229,14 @@ void __cdecl FUN_00504b50(int param_1, int param_2, float *param_3,
         *(float *)(param_1 + 0x6c) = rotU;
         *(float *)(param_1 + 0x70) = (float)WorldTime * -0.0005f;
         *(char *)(DAT_05828d58 + effectType * 0xbc + 0x88) = 0;
-        FUN_00441e00(this_, 2, *(float *)(param_1 + 0x168), *(float *)(param_1 + 100),
+        FUN_00441e00(this_, 2, *(float *)(param_1 + 0x168), *(int *)(param_1 + 100),
                      *(float *)(param_1 + 0x68), *(float *)(param_1 + 0x6c),
                      *(float *)(param_1 + 0x70), *(float *)(param_1 + 0x58), 1170);
         *(char *)(DAT_05828d58 + effectType * 0xbc + 0x88) = -1;
         *(float *)((int)this_ + 0x48) = 1.0f;
         *(float *)((int)this_ + 0x4c) = 1.0f;
         *(float *)((int)this_ + 0x50) = 1.0f;
-        FUN_00441e00(this_, 2, *(float *)(param_1 + 0x168), *(float *)(param_1 + 100),
+        FUN_00441e00(this_, 2, *(float *)(param_1 + 0x168), *(int *)(param_1 + 100),
                      *(float *)(param_1 + 0x68), *(float *)(param_1 + 0x6c),
                      *(float *)(param_1 + 0x70), *(float *)(param_1 + 0x58), -1);
         return;
@@ -247,7 +247,7 @@ void __cdecl FUN_00504b50(int param_1, int param_2, float *param_3,
         *(float *)((int)this_ + 0x4c) = 0.8f;
         *(float *)((int)this_ + 0x50) = 0.8f;
         pulse = sinf((float)WorldTime * 0.002f) * 0.3f + 0.7f;
-        FUN_00441e00(this_, 0x42, 1.0f, 0.0f, pulse,
+        FUN_00441e00(this_, 0x42, 1.0f, 0, pulse,
                      *(float *)(param_1 + 0x6c), *(float *)(param_1 + 0x70),
                      *(float *)(param_1 + 0x58), -1);
         return;
@@ -507,7 +507,7 @@ LAB_after_render: ;
 
         // flags=0x42 (int flags, NOT a pointer!), f1=1.0, then anim data
         FUN_00441e00(this_, 0x42, 1.0f,
-                     *(float*)(param_1 + 100),
+                     *(int *)(param_1 + 100),
                      *(float*)(param_1 + 0x68),
                      *(float*)(param_1 + 0x6c),
                      *(float*)(param_1 + 0x70),
