+++ src/Render/Entity_Render.cpp
@@ -594,7 +594,7 @@ void __cdecl FUN_00505970(void *param_1,void *param_2_v,int param_3,char param_4
   }
   FUN_004404e0(param_1,(int)puVar1,(float *)(param_2 + 0x118),(float *)(param_2 + 0x124),
                (float *)(param_2 + 0x130),param_4);
-  FUN_00441e00(param_1,(uint)param_3,*(float *)(param_2 + 0x168),*(float *)(param_2 + 100),
+  FUN_00441e00(param_1,(uint)param_3,*(float *)(param_2 + 0x168),*(int *)(param_2 + 100),
                *(float *)(param_2 + 0x68),*(float *)(param_2 + 0x6c),*(float *)(param_2 + 0x70),
                *(float *)(param_2 + 0x58),0xffffffff);
   _DAT_005597c8 = 1.0f;   // reset LOD scale
